refactor(runtime): extract turn/outcome.rs into agent/turn — #7415
